Conflicts:
	.gradle/2.14.1/taskArtifacts/cache.properties
	build.gradle
	eMobilePOSAndroid/eMobilePOSAndroid-eMobilePOSAndroid.iml
	eMobilePOSAndroid/libs/mips/libRBA_SDK.so
	eMobilePOSAndroid/libs/mips/libgnustl_shared.so
	eMobilePOSAndroid/src/main/java/com/android/database/OrdersHandler.java
	eMobilePOSAndroid/src/main/java/com/android/emobilepos/history/details/HistoryTransactionDetails_FA.java
	eMobilePOSAndroid/src/main/java/com/android/emobilepos/mainmenu/MainMenu_FA.java
	eMobilePOSAndroid/src/main/res/values-es/strings.xml
	eMobilePOSAndroid/src/main/res/values/strings.xml

+++ eMobilePOSAndroid/src/main/java/drivers/EMSDeviceDriver.java
@@ -721,7 +721,7 @@ public class EMSDeviceDriver {
                         payWithLoyalty = true;
                     }
                     totalItemstQty += TextUtils.isEmpty(orderProducts.get(i).getOrdprod_qty()) ? 0
-                            : Integer.parseInt(orderProducts.get(i).getOrdprod_qty());
+                            : Double.parseDouble(orderProducts.get(i).getOrdprod_qty());
                     sb.append(orderProducts.get(i).getOrdprod_name()).append("-").append(orderProducts.get(i).getOrdprod_desc())
                             .append("\n");
